Support automatic version update via Docker
If an existing Joomla installation in the /var/www/html volume is found,
compare installed Joomla version with the one in this Docker Image
and run an upgrade if Minor- and/or Patch-Version are newer.

This is done by overwriting the existing Joomla Core files in the volume
with the newly downloaded ones from the Docker Image, running the
"Post-Manual Update Script" from https://gist.github.com/mbabker/d7bfb4e1e2fbc6b7815a733607f89281
afterwards and removing the 'installation' directory again.