Added php intl extension and dependencies
This extension is now recommended by the Joomla! update component when upgrading to version 4.0.3, so I seized the opportunity and updated the Dockerfile.template for version 4 to include it and its dependencies by default. However, these changes have not yet been verified to work properly due to a lack of time, so some minor adjustments may be required.