Use `mysqli_report(MYSQLI_REPORT_OFF)` in `makedb.php`
Using PHP 8.1, the `makedb.php` script fails if the connection cannot be
established at the first try.  This is because according to
https://www.php.net/manual/en/mysqli-driver.report-mode.php, the default mysqli
reporting mode is now `MYSQLI_REPORT_ERROR | MYSQLI_REPORT_STRICT` while it was
`MYSQLI_REPORT_OFF` before.  The change makes `mysqli::__construct()` throw an
exception, which is not caught.  This commit restores the old behavior.